requirements.txt
django-redis==4.3.0
redis==2.10.5
hiredis==0.2.0
